fix: poll directories to prevent TS6307 (#454)
Webpack watcher notifies only about changes of files that are in the compilation dependencies graph. This means that it will not notice if we create a new file that is not missing. This is a difference between webpack and typescript watch mechanism that can lead to errors like TS6307. This fix polls directories content on getReport call and compare it with the previous content.